feat: error handling for AlreadyExists

+++ src/add.rs
@@ -1,5 +1,6 @@
 use std::fs::File;
-use std::io::Result;
+use std::io::{Result, Error, ErrorKind};
+use std::path::Path;
 use structopt::StructOpt;
 
 #[derive(StructOpt)]
@@ -10,6 +11,11 @@ pub struct AddOptions {
 
 pub fn add(folder: &str, args: &AddOptions) -> Result<()> {
     let path = folder.to_string() + &args.file;
-    File::create(path).expect("failed to create file");
-    Ok(())
+    // error handling for AlreadyExists.
+    if Path::new(&path).exists() {
+        Err(Error::new(ErrorKind::AlreadyExists, "already existed this file."))
+    } else {
+        File::create(path).expect("failed to create file");
+        Ok(())
+    }
 }
